refactor: rename and restructure service-account.ts and connect.ts
